style: mypy
