+++ dynamicio/config/pydantic/io_resources.py
@@ -10,8 +10,6 @@ import pydantic
 
 import dynamicio.config.pydantic.table_schema as table_spec
 
-# import dynamicio.config.pydantic.config
-
 
 @enum.unique
 class DataBackendType(str, enum.Enum):
@@ -40,10 +38,61 @@ class FileType(str, enum.Enum):
     txt = "txt"
 
 
+class IOBinding(pydantic.BaseModel):
+    """A binding for a single i/o object"""
+
+    name: str = pydantic.Field(alias="__binding_name__")
+    environments: typing.Mapping[str, "IOEnvironment"]
+    dynamicio_schema: typing.Union[table_spec.DataframeSchema, None] = pydantic.Field(default=None, alias="schema")
+
+    def get_binding_for_environment(self, environment: str) -> "IOEnvironment":
+        """Fetch the IOEnvironment spec for the name provided."""
+        return self.environments[environment]
+
+    @pydantic.validator("environments", pre=True, always=True)
+    def pick_correct_env_cls(cls, value, values, config, field):
+        """This pre-validator picks an appropriate IOEnvironment subclass for the `data_backend_type`"""
+        if not isinstance(value, typing.Mapping):
+            raise ValueError(f"Environments input should be a dict. Got {value!r} instead.")
+        config_cls_overrides = {
+            DataBackendType.local: LocalDataEnvironment,
+            DataBackendType.local_batch: LocalBatchDataEnvironment,
+            DataBackendType.s3: S3DataEnvironment,
+            DataBackendType.s3_file: S3DataEnvironment,
+            DataBackendType.s3_path_prefix: S3PathPrefixEnvironment,
+            DataBackendType.kafka: KafkaDataEnvironment,
+            DataBackendType.postgres: PostgresDataEnvironment,
+        }
+        out_dict = {}
+        for (env_name, env_data) in value.items():
+            base_obj: IOEnvironment = field.type_(**env_data)
+            override_cls = config_cls_overrides.get(base_obj.data_backend_type)
+            if override_cls:
+                use_obj = override_cls(**env_data)
+            else:
+                use_obj = base_obj
+            out_dict[env_name] = use_obj
+        return out_dict
+
+    @pydantic.root_validator(pre=True)
+    def _preprocess_raw_config(cls, values):
+        if not isinstance(values, typing.Mapping):
+            raise ValueError(f"IOBinding must be a dict at the top level. (got {values!r} instead)")
+        remapped_value = {"environments": {}}
+        for (key, value) in values.items():
+            if key in ("__binding_name__", "schema"):
+                # Passthrough params
+                remapped_value[key] = value
+            else:
+                # Assuming an environment config
+                remapped_value["environments"][key] = value
+        return remapped_value
+
+
 class IOEnvironment(pydantic.BaseModel):
     """A section specifiing an data source backed by a particular data backend"""
 
-    _parent: typing.Optional["dynamicio.config.pydantic.config.IOBinding"] = None  # noqa: F821
+    _parent: typing.Optional[IOBinding] = None  # noqa: F821
     options: typing.Mapping = pydantic.Field(default_factory=dict)
     data_backend_type: DataBackendType = pydantic.Field(alias="type", const=None)
 
@@ -56,10 +105,10 @@ class IOEnvironment(pydantic.BaseModel):
     def dynamicio_schema(self) -> typing.Union[table_spec.DataframeSchema, None]:
         """Returns tabular data structure definition for the data source (if available)"""
         if not self._parent:
-            return None
+            raise Exception("Parent field is not set.")
         return self._parent.dynamicio_schema
 
-    def set_parent(self, parent: "dynamicio.config.pydantic.config.IOBinding"):  # noqa: F821
+    def set_parent(self, parent: IOBinding):  # noqa: F821
         """Helper method to set parent config object."""
         assert self._parent is None
         self._parent = parent
@@ -167,3 +216,6 @@ class PostgresDataEnvironment(IOEnvironment):
     """Parent section for postgres data source."""
 
     postgres: PostgresDataSubSection
+
+
+IOBinding.update_forward_refs()
